Stream dependencies (#1432)
* Create input directory

Ensure that the input directory is created when building container
images.

* Add support for streaming container dependencies

Container dependencies (inputs) no longer require the use of volume
mounts. They are streamed into the container using the Docker API.

Dependencies must first be added to a TAR file before they can be copied
into the container. This is handled by the container_dependencies
module.

Hyper has been chosen as the HTTP client for interacting with the Docker
API. Although a synchronous HTTP client would have been more convenient,
none were suitable for streaming directly to the container. As a result,
the interface between synchronous and asynchronous code is handled by
the ContainerDependencyWriter structure.

The NullSink structure has been added to facilitate calculating the size
of the ChunkReader stream. This is a requirement of the TAR builder when
adding a dependency to it.