FEAT-3012: Retrieve json from published data science dcr (#2548)
Provide the ability to retrieve the JSON configuration of a Data Science DCR via the SDK. Other DCR types are not currently supported. The steps required to retrieve the JSON follow those taken in the UI code whereby we:

   - Retrieve the data room from the driver
   - Verify the data room
   - Upgrade it to the latest version
   - Return it as JSON

Support has also been provided for retrieval of the DCR kind via the client.