FEAT-3034: 11.1% Rounding (#2572)
Replace the rounding function with a function to add noise (from a Laplace distribution). This makes the insights tab more useful for low match sizes.

With the rounding function (rounding to the nearest 100) it was noted that a customer saw multiples of 11.1% in the "Share in Overlap" column - thereby not providing much discernible information. The cause of this was due to “the common scenario of 900 matched users, so all segments must be smaller (100, 200, 300, 400), so all % are multiples of 1/9".

By adding a random amount of noise which is typically smaller than 100, we improve the granularity of the outputs.