FEAT-4633: Limit the impact of noise addition and ensure the noise is the same for all participants (#3780)
Make use of the RESULT_HASH_HASH environment variable to seed the random number generator when adding noise. This ensures that multiple participants will have the same amount of noise added so long as nothing in the compute graph changes.

Additionally limit the amount of noise that is added to a value to be within a +/- 20% bounds. The final result is rounded which can lead to it being slightly outside these bounds for small numbers. Nonetheless, the amount of noise added is contained.