FEAT-4858: Permutive fixes (#3888)
* FEAT-4858: Ensure Permutive upload content type is set to gzip (#3831)

When uploading the Permutive file to a GCS bucket it is in a gzip format. In order for the format to be correctly interpreted in GCS the content-header must be set to application/gzip when creating the multipart upload in the S3 sink worker.

To maintain backwards compatibility the specified content type is made optional and is only currently specified in the Permutive upload case.

The documentation has also been updated to specify that a .gz file extension is required when specifying the object name. The accompanying images have also been updated to reflect this.

* FEAT-4831: Ensure the output of the Permutive worker is in a gzip format (#3801)

Note the following from the Permutive docs:

"The file must be a tab-separated file. For efficiency, data files should be uploaded in compressed format. Please gzip and compress your files prior to uploading. Post compression, file names should end with a ".gz" extension. If the files end with a ".gzip" extension, the import will not work."

This requires that the object name provided by the user has a .gz file extension.

Also changed the Permutive worker so that we don't get an extra newline at the end of the file.

* FEAT-4779: Add bucket and object details to Permutive config (#3781)

Display the bucket name and object name in the UI for the Permutive export connector.